Refactor Response object
