+++ tests/test_response.py
@@ -16,8 +16,9 @@
 
 from __future__ import division
 
-import unittest
 import concurrent.futures
+import itertools
+import unittest
 
 from collections import OrderedDict
 
@@ -25,274 +26,83 @@ import numpy as np
 
 import dimod
 
-try:
-    import pandas as pd
-    _pandas = True
-except ImportError:
-    _pandas = False
-
 
 class TestResponse(unittest.TestCase):
+    def test__repr__(self):
 
-    def test_instantiation(self):
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0],
-                                   [0, 0, 0, 0],
-                                   [1, 1, 1, 1]])
-        energies = np.asarray([2, 2, 0, 4], dtype=float)
+        alpha = list('abcdefghijklmnopqrstuvwxyz')
 
-        response = dimod.Response(samples_matrix, {'energy': energies}, dimod.BINARY)
+        num_variables = 3  # needs to be small enough to not snip the matrix when printed
+        num_samples = 2 ** num_variables
 
-        np.testing.assert_equal(samples_matrix, response.record.sample)
-        np.testing.assert_allclose(energies, response.record.energy)
+        samples = list(itertools.product((-1, 1), repeat=num_variables))
+        energies = [0.0]*num_samples
+        num_occurrences = [1] * num_samples
+        record = dimod.response.data_struct_array(samples, energy=energies, num_occurrences=num_occurrences)
 
-    def test_data_vectors_are_arrays(self):
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0],
-                                   [0, 0, 0, 0],
-                                   [1, 1, 1, 1]])
-        energies = [2, 2, 0, 4]
-        num_occurrences = [1, 1, 2, 1]
-        objects = [object() for __ in range(4)]
+        labels = alpha[:num_variables]
 
-        data_vectors = {'energy': energies, 'occurences': num_occurrences, 'objects': objects}
+        resp = dimod.Response(record, labels, {}, dimod.SPIN)
 
-        response = dimod.Response(samples_matrix, data_vectors, dimod.BINARY)
+        from dimod import Response
+        from numpy import rec
+        new_resp = eval(resp.__repr__())
 
-        self.assertEqual(len(response.record.dtype.fields), 4)
+        np.testing.assert_array_equal(resp.record, new_resp.record)
 
-        for key in data_vectors:
-            self.assertIn(key, response.record.dtype.fields)
+    def test_from_samples_dicts(self):
+        bqm = dimod.BinaryQuadraticModel({'a': .1}, {('a', 'b'): -1}, 0.0, dimod.SPIN)
+        samples = [{'a': -1, 'b': +1}, {'a': +1, 'b': -1},
+                   {'a': +1, 'b': +1}, {'a': -1, 'b': -1}]
+        energies = [bqm.energy(sample) for sample in samples]
 
-            vector = response.record[key]
+        resp = dimod.Response.from_samples(samples, {'energy': energies}, {}, dimod.SPIN)
 
-            self.assertIsInstance(vector, np.ndarray)
+    def test_from_future_default(self):
 
-            self.assertEqual(vector.shape, (4,))  # in this case they will all be 1D
+        future = concurrent.futures.Future()
 
-    def test_data_vectors_wrong_length(self):
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0],
-                                   [0, 0, 0, 0],
-                                   [1, 1, 1, 1]])
-        energies = [2, 2, 0, 4]
-        num_occurrences = [1, 1, 2, 1, 1]
-        objects = [object() for __ in range(4)]
+        response = dimod.Response.from_future(future)
 
-        data_vectors = {'energy': energies, 'occurences': num_occurrences, 'objects': objects}
+        self.assertIsInstance(response, dimod.Response)
+        self.assertFalse(hasattr(response, '_record'))  # should not have a record yet
 
-        with self.assertRaises(ValueError):
-            response = dimod.Response(samples_matrix, data_vectors, dimod.BINARY)
+        self.assertFalse(response.done())
 
-    def test_data_vectors_not_array_like(self):
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0],
-                                   [0, 0, 0, 0],
-                                   [1, 1, 1, 1]])
-        energies = [2, 2, 0, 4]
-        num_occurrences = 'hi there'
-        objects = [object() for __ in range(4)]
+        # make future return a Response
+        future.set_result(dimod.Response.from_samples([-1, -1, 1], {'energy': [.5]}, {'another_field': .5}, dimod.SPIN))
 
-        data_vectors = {'energy': energies, 'occurences': num_occurrences, 'objects': objects}
+        self.assertTrue(response.done())
 
-        with self.assertRaises(ValueError):
-            response = dimod.Response(samples_matrix, data_vectors, dimod.BINARY)
-
-    def test_samples_num_limited(self):
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0],
-                                   [0, 0, 0, 0],
-                                   [1, 1, 1, 1]])
-        energies = [2, 2, 0, 4]
-        num_occurrences = [1, 1, 2, 1]
-        objects = [object() for __ in range(4)]
-
-        data_vectors = {'energy': energies, 'occurences': num_occurrences, 'objects': objects}
-
-        response = dimod.Response(samples_matrix, data_vectors, dimod.BINARY)
-
-        samples_list = list(response.samples())
-
-        self.assertEqual(len(samples_list), 4)
-
-        shortened_samples_list = list(response.samples(3))
-
-        self.assertEqual(len(shortened_samples_list), 3)
-        self.assertEqual(shortened_samples_list, samples_list[0:3])
-
-    def test_instantiation_without_energy(self):
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0],
-                                   [0, 0, 0, 0],
-                                   [1, 1, 1, 1]])
-
-        with self.assertRaises(ValueError):
-            dimod.Response(samples_matrix, {}, dimod.BINARY)
-
-    def test_from_matrix(self):
-        # binary matrix
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0]])
-        energies = np.asarray([0, 1], dtype=float)
-
-        response = dimod.Response.from_matrix(samples_matrix, {'energy': energies}, vartype=dimod.BINARY)
-
-        self.assertTrue(np.all(samples_matrix == response.record.sample))
-        self.assertEqual(response.record.sample.dtype, np.int8)
-        self.assertIs(response.vartype, dimod.BINARY)
-
-        # spin array
-        samples_matrix = np.asarray([[-1, +1, -1, +1],
-                                     [+1, -1, +1, -1]])
-        energies = np.asarray([-.5, 1], dtype=float)
-
-        response = dimod.Response.from_matrix(samples_matrix, {'energy': energies}, dimod.SPIN)
-
-        self.assertTrue(np.all(samples_matrix == response.record.sample))
-        self.assertEqual(response.record.sample.dtype, np.int8)
+        # accessing response.record should resolve the future
+        np.testing.assert_array_equal(response.record.sample,
+                                      np.array([[-1, -1, 1]]))
+        np.testing.assert_array_equal(response.record.energy,
+                                      np.array([.5]))
+        np.testing.assert_array_equal(response.record.num_occurrences,
+                                      np.array([1]))
+        self.assertEqual(response.info, {'another_field': .5})
         self.assertIs(response.vartype, dimod.SPIN)
+        self.assertEqual(response.labels, [0, 1, 2])
 
-        # with array type
-        samples_matrix = np.array([[0, 1, 0, 1],
-                                   [1, 0, 1, 0]])
-        energies = np.asarray([0, 1], dtype=float)
+    def test__iter__(self):
+        samples = [{'a': 0, 'b': 1}, {'a': 1, 'b': 0},
+                   {'a': 1, 'b': 1}, {'a': 0, 'b': 0}]
+        energies = [0, 1, 2, 3]
+        response = dimod.Response.from_samples(samples, {'energy': energies}, {}, dimod.BINARY)
 
-        response = dimod.Response.from_matrix(samples_matrix, {'energy': energies}, vartype=dimod.BINARY)
+        for s0, s1 in zip(response, samples):
+            self.assertEqual(s0, s1)
 
-        self.assertTrue(np.all(samples_matrix == response.record.sample))
-        self.assertEqual(response.record.sample.dtype, np.int8)
-        self.assertIs(response.vartype, dimod.BINARY)
-
-    def test_from_dicts(self):
-        """Typical use"""
-
-        samples = [{'a': -1, 'b': +1},
-                   {'a': -1, 'b': -1}]
-        energies = [sum(sample.values()) for sample in samples]
-
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
-
-        self.assertTrue(np.all(response.record.sample == np.asarray([[-1, +1], [-1, -1]])))
-        self.assertEqual(response.record.sample.dtype, np.int8)
-        self.assertIs(response.vartype, dimod.SPIN)
-        self.assertEqual(response.variable_labels, ['a', 'b'])  # sortable in py2 and py3
-
-    def test_from_dicts_unsortable_labels(self):
-        """Python3 cannot sort unlike types"""
-        samples = [OrderedDict([('a', -1), (0, +1)]),  # ordered when cast to a list
-                   OrderedDict([('a', -1), (0, -1)])]
-        energies = [sum(sample.values()) for sample in samples]
-
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
-
-        if response.variable_labels == [0, 'a']:
-            np.testing.assert_equal(response.record.sample, np.array([[+1, -1], [-1, -1]]))
-        else:
-            self.assertEqual(response.variable_labels, ['a', 0])
-            np.testing.assert_equal(response.record.sample, np.array([[-1, +1], [-1, -1]]))
-        self.assertEqual(response.record.sample.dtype, np.int8)
-        self.assertIs(response.vartype, dimod.SPIN)
-
-    def test_from_dicts_unlike_labels(self):
+    def test_from_samples_unlike_labels(self):
         samples = [{'a': -1, 'b': +1},
                    {'a': -1, 'c': -1}]
         energies = [sum(sample.values()) for sample in samples]
 
         with self.assertRaises(ValueError):
-            dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
+            dimod.Response.from_samples(samples, {'energy': energies}, {}, dimod.SPIN)
 
-    @unittest.skipUnless(_pandas, "no pandas installed")
-    def test_from_pandas(self):
-        samples_df = pd.DataFrame([[0, 1, 0], [1, 1, 1]], columns=['a', 'b', 'c'])
-
-        response = dimod.Response.from_pandas(samples_df, {'energy': [-1, 1]}, dimod.BINARY)
-
-    def test_update(self):
-        samples = [{'a': -1, 'b': +1},
-                   {'a': -1, 'b': -1}]
-        energies = [sum(sample.values()) for sample in samples]
-
-        response0 = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
-        response1 = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
-
-        response0.update(response1)
-
-        self.assertEqual(len(response0), 4)
-        for field in response0.record.dtype.fields:
-            self.assertEqual(len(response0.record[field]), 4)
-
-    def test_update_energy(self):
-
-        def _spingen():
-            while True:
-                yield -1
-                yield +1
-                yield -1
-        spingen = _spingen()
-
-        h = {v: v * .2 for v in range(8)}
-        J = {(u, u + 1): .3 * u * (u + 1) for u in range(7)}
-
-        samples = [{v: next(spingen) for v in range(8)}
-                   for __ in range(10)]
-        energies = [dimod.ising_energy(sample, h, J) for sample in samples]
-
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, vartype=dimod.SPIN)
-
-        # first make sure that all the energies are good
-        for sample, energy in response.data(['sample', 'energy']):
-            self.assertAlmostEqual(energy, dimod.ising_energy(sample, h, J))
-
-        # now do a bunch of updates
-        for __ in range(10):
-            samples = [{v: next(spingen) for v in range(8)}
-                       for __ in range(10)]
-            energies = [dimod.ising_energy(sample, h, J) for sample in samples]
-
-            new_response = dimod.Response.from_dicts(samples, {'energy': energies}, vartype=dimod.SPIN)
-
-            response.update(new_response)
-
-        self.assertEqual(len(response), 110)
-
-        # and check again
-        for sample, energy in response.data(['sample', 'energy']):
-            self.assertAlmostEqual(energy, dimod.ising_energy(sample, h, J))
-
-    def test_from_futures(self):
-        def _futures():
-            for __ in range(2):
-                future = concurrent.futures.Future()
-
-                future.set_result({'samples': [-1, -1, 1], 'energy': [.5]})
-
-                yield future
-
-        response = dimod.Response.from_futures(_futures(), vartype=dimod.SPIN, num_variables=3)
-
-        self.assertTrue(response.done())
-
-        matrix = response.record.sample
-
-        np.testing.assert_array_equal(matrix, np.array([[-1, -1, 1], [-1, -1, 1]]))
-
-    def test_from_futures_column_subset(self):
-        def _futures():
-            for __ in range(2):
-                future = concurrent.futures.Future()
-
-                future.set_result({'samples': [-1, -1, 1], 'energy': [.5]})
-
-                yield future
-
-        response = dimod.Response.from_futures(_futures(), vartype=dimod.SPIN, num_variables=2,
-                                               active_variables=[0, 2])
-
-        matrix = response.record.sample
-
-        np.testing.assert_equal(matrix, np.array([[-1, 1], [-1, 1]]))
-
-    def test_from_futures_typical(self):
+    def test_from_future_typical(self):
         result = {'occurrences': [1],
                   'active_variables': [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
                                        18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32,
@@ -325,15 +135,13 @@ class TestResponse(unittest.TestCase):
 
         future = concurrent.futures.Future()
 
-        response = dimod.Response.from_futures([future],
-                                               vartype=dimod.SPIN,
-                                               num_variables=result['num_variables'],
-                                               samples_key='samples',
-                                               data_vector_keys={'energies': 'energy',
-                                                                 'occurrences': 'occurrences'},
-                                               info_keys=['timing'],
-                                               variable_labels=result['active_variables'],
-                                               active_variables=result['active_variables'])
+        def result_to_response(future):
+            result = future.result()
+            return dimod.Response.from_samples(result['solutions'],
+                                               {'energy': result['energies'], 'num_occurrences': result['occurrences']},
+                                               {}, dimod.SPIN)
+
+        response = dimod.Response.from_future(future, result_hook=result_to_response)
 
         future.set_result(result)
 
@@ -341,67 +149,13 @@ class TestResponse(unittest.TestCase):
 
         np.testing.assert_equal(matrix, result['samples'])
 
-    def test_from_futures_extra_keys(self):
-        def _futures():
-            for __ in range(2):
-                future = concurrent.futures.Future()
-
-                future.set_result({'samples': [-1, -1, 1], 'energy': [.5]})
-
-                yield future
-
-        response = dimod.Response.from_futures(_futures(), vartype=dimod.SPIN, num_variables=2,
-                                               active_variables=[0, 2],
-                                               data_vector_keys={'energy', 'other'},
-                                               info_keys=['other'])
-
-        matrix = response.record.sample
-
-        np.testing.assert_equal(matrix, np.array([[-1, 1], [-1, 1]]))
-
-        #
-
-        response = dimod.Response.from_futures(_futures(), vartype=dimod.SPIN, num_variables=2,
-                                               active_variables=[0, 2],
-                                               data_vector_keys={'energy', 'other'},
-                                               info_keys=['other'],
-                                               ignore_extra_keys=False)
-
-        with self.assertRaises(ValueError):
-            matrix = response.record.sample
-
-        #
-
-        response = dimod.Response.from_futures(_futures(), vartype=dimod.SPIN, num_variables=2,
-                                               active_variables=[0, 2],
-                                               data_vector_keys={'energy'},
-                                               info_keys=['other'],
-                                               ignore_extra_keys=False)
-
-        with self.assertRaises(ValueError):
-            matrix = response.record.sample
-
-    def test_empty(self):
-        response = dimod.Response.empty(dimod.SPIN)
-        self.assertFalse(response)
-        self.assertIs(response.vartype, dimod.SPIN)
-
-    def test__iter__(self):
-        samples = [{'a': 0, 'b': 1}, {'a': 1, 'b': 0},
-                   {'a': 1, 'b': 1}, {'a': 0, 'b': 0}]
-        energies = [0, 1, 2, 3]
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.BINARY)
-
-        for s0, s1 in zip(response, samples):
-            self.assertEqual(s0, s1)
-
     def test_data_docstrings(self):
         samples = [{'a': 0, 'b': 1}, {'a': 1, 'b': 0}, {'a': 0, 'b': 0}]
         energies = [-1, 0, 1]
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, vartype=dimod.BINARY)
+        response = dimod.Response.from_samples(samples, {'energy': energies}, {}, vartype=dimod.BINARY)
 
         for datum in response.data():
-            sample, energy = datum  # this should work
+            sample, energy, occ = datum  # this should work
             self.assertEqual(sample, datum.sample)  # named tuple
             self.assertEqual(energy, datum.energy)  # named tuple
             self.assertIn(datum.sample, samples)
@@ -413,8 +167,8 @@ class TestResponse(unittest.TestCase):
         samples = [{'a': -1, 'b': +1}, {'a': +1, 'b': -1}, {'a': -1, 'b': -1}]
         energies = [-1.0, 0.0, 1.0]
         num_spin_up = [1, 1, 0]
-        response = dimod.Response.from_dicts(samples, {'energy': energies, 'num_spin_up': num_spin_up},
-                                             vartype=dimod.SPIN)
+        response = dimod.Response.from_samples(samples, {'energy': energies, 'num_spin_up': num_spin_up},
+                                               {}, vartype=dimod.SPIN)
         for datum in response.data(['num_spin_up', 'energy']):
             nsu, energy = datum
             self.assertEqual(nsu, datum.num_spin_up)
@@ -425,7 +179,7 @@ class TestResponse(unittest.TestCase):
         samples = [{'a': -1, 'b': +1}, {'a': +1, 'b': -1},
                    {'a': +1, 'b': +1}, {'a': -1, 'b': -1}]
         energies = [bqm.energy(sample) for sample in samples]
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
+        response = dimod.Response.from_samples(samples, {'energy': energies}, {}, dimod.SPIN)
 
         for sample, energy in response.data(['sample', 'energy']):
             self.assertAlmostEqual(bqm.energy(sample), energy)
@@ -446,7 +200,7 @@ class TestResponse(unittest.TestCase):
             self.assertAlmostEqual(bqm.spin.energy(sample), energy)
 
         # finally change with an offset
-        response.change_vartype(dimod.BINARY, {'energy': 1})
+        response.change_vartype(dimod.BINARY, 1)
         for sample, energy in response.data(['sample', 'energy']):
             self.assertAlmostEqual(bqm.binary.energy(sample) + 1, energy)
 
@@ -455,7 +209,7 @@ class TestResponse(unittest.TestCase):
         samples = [{'a': -1, 'b': +1}, {'a': +1, 'b': -1},
                    {'a': +1, 'b': +1}, {'a': -1, 'b': -1}]
         energies = [bqm.energy(sample) for sample in samples]
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
+        response = dimod.Response.from_samples(samples, {'energy': energies}, {}, dimod.SPIN)
 
         for sample, energy in response.data(['sample', 'energy']):
             self.assertAlmostEqual(bqm.energy(sample), energy)
@@ -472,7 +226,7 @@ class TestResponse(unittest.TestCase):
         samples = [{'a': -1, 'b': +1}, {'a': +1, 'b': -1},
                    {'a': -1, 'b': +1}, {'a': -1, 'b': -1}]
         energies = [bqm.energy(sample) for sample in samples]
-        response = dimod.Response.from_dicts(samples, {'energy': energies}, dimod.SPIN)
+        response = dimod.Response.from_samples(samples, {'energy': energies}, {}, dimod.SPIN)
 
         new_response = response.relabel_variables({'a': 0, 'b': 1}, inplace=False)
 
@@ -484,14 +238,13 @@ class TestResponse(unittest.TestCase):
             self.assertEqual(set(sample), {0, 1})
 
     def test_relabel_docstring(self):
-        response = dimod.Response.from_dicts([{'a': -1}, {'a': +1}], {'energy': [-1, 1]}, dimod.SPIN)
-        response = dimod.Response.from_dicts([{'a': -1}, {'a': +1}], {'energy': [-1, 1]}, dimod.SPIN)
+        response = dimod.Response.from_samples([{'a': -1}, {'a': +1}], {'energy': [-1, 1]}, {}, dimod.SPIN)
         new_response = response.relabel_variables({'a': 0}, inplace=False)
 
     def test_partial_relabel_inplace(self):
         mapping = {0: '3', 1: 4, 2: 5, 3: 6, 4: 7, 5: '1', 6: '2', 7: '0'}
 
-        response = dimod.Response.from_matrix([[-1, +1, -1, +1, -1, +1, -1, +1]], {'energy': [-1]}, dimod.SPIN)
+        response = dimod.Response.from_samples([[-1, +1, -1, +1, -1, +1, -1, +1]], {'energy': [-1]}, {}, dimod.SPIN)
 
         new_response = response.relabel_variables(mapping, inplace=False)
 
@@ -506,8 +259,8 @@ class TestResponse(unittest.TestCase):
 
         mapping = {0: '3', 1: 4, 2: 5, 3: 6, 4: 7, 5: '1', 6: '2', 7: '0'}
 
-        response = dimod.Response.from_matrix([[-1, +1, -1, +1, -1, +1, -1, +1]], {'energy': [-1]}, dimod.SPIN)
-        response2 = dimod.Response.from_matrix([[-1, +1, -1, +1, -1, +1, -1, +1]], {'energy': [-1]}, dimod.SPIN)
+        response = dimod.Response.from_samples([[-1, +1, -1, +1, -1, +1, -1, +1]], {'energy': [-1]}, {}, dimod.SPIN)
+        response2 = dimod.Response.from_samples([[-1, +1, -1, +1, -1, +1, -1, +1]], {'energy': [-1]}, {}, dimod.SPIN)
 
         response.relabel_variables(mapping, inplace=True)
 
@@ -525,7 +278,7 @@ class TestSamplesStructuredArray(unittest.TestCase):
 
         self.assertEqual(data.shape, (0,))
 
-        self.assertEqual(len(data.dtype.fields), 2)
+        self.assertEqual(len(data.dtype.fields), 3)
         self.assertIn('sample', data.dtype.fields)
         self.assertIn('energy', data.dtype.fields)
 
@@ -534,7 +287,7 @@ class TestSamplesStructuredArray(unittest.TestCase):
 
         self.assertEqual(data.shape, (1,))
 
-        self.assertEqual(len(data.dtype.fields), 2)
+        self.assertEqual(len(data.dtype.fields), 3)
         self.assertIn('sample', data.dtype.fields)
         self.assertIn('energy', data.dtype.fields)
 
@@ -543,7 +296,7 @@ class TestSamplesStructuredArray(unittest.TestCase):
 
         self.assertEqual(data.shape, (1,))
 
-        self.assertEqual(len(data.dtype.fields), 2)
+        self.assertEqual(len(data.dtype.fields), 3)
         self.assertIn('sample', data.dtype.fields)
         self.assertIn('energy', data.dtype.fields)
 
@@ -552,29 +305,31 @@ class TestSamplesStructuredArray(unittest.TestCase):
 
         self.assertEqual(data.shape, (2,))
 
-        self.assertEqual(len(data.dtype.fields), 2)
+        self.assertEqual(len(data.dtype.fields), 3)
         self.assertIn('sample', data.dtype.fields)
         self.assertIn('energy', data.dtype.fields)
 
     def test_extra_data_vector(self):
-        data = dimod.response.data_struct_array([[-1, +1, -1], [+1, -1, +1]], energy=[1.5, 4.5], occurrences=np.asarray([1, 2]))
+        data = dimod.response.data_struct_array([[-1, +1, -1], [+1, -1, +1]], energy=[1.5, 4.5],
+                                                num_occurrences=np.asarray([1, 2]))
 
         self.assertEqual(data.shape, (2,))
 
         self.assertEqual(len(data.dtype.fields), 3)
         self.assertIn('sample', data.dtype.fields)
         self.assertIn('energy', data.dtype.fields)
-        self.assertIn('occurrences', data.dtype.fields)
+        self.assertIn('num_occurrences', data.dtype.fields)
 
     def test_data_vector_higher_dimension(self):
-        data = dimod.response.data_struct_array([[-1, +1, -1], [+1, -1, +1]], energy=[1.5, 4.5], occurrences=[[0, 1], [1, 2]])
+        data = dimod.response.data_struct_array([[-1, +1, -1], [+1, -1, +1]], energy=[1.5, 4.5],
+                                                num_occurrences=[[0, 1], [1, 2]])
 
         self.assertEqual(data.shape, (2,))
 
         self.assertEqual(len(data.dtype.fields), 3)
         self.assertIn('sample', data.dtype.fields)
         self.assertIn('energy', data.dtype.fields)
-        self.assertIn('occurrences', data.dtype.fields)
+        self.assertIn('num_occurrences', data.dtype.fields)
 
     def test_mismatched_vector_samples_rows(self):
         with self.assertRaises(ValueError):
